feat: Add detection of images in child of a block + Add DEBUG environment
When images was in a child of a page block (for example a column_list, images wasn't detected.
Now, if block have has_children property to True, we follow the child block (without level limit)

Also add a DEBUG environment variable. Set to False if not defined or explicitly defined to False.
Need to be explicitly define to True to enable debug logs